Move array of housemates into a different file for repeated use
