+++ src/components/Voting.jsx
@@ -1,31 +1,6 @@
 import React, { useState } from "react";
 import Housemate from "./Housemate";
-
-const staticHousemates = [
-  {
-    name: "Dorathy",
-    picture: "https://dailypost.ng/wp-content/uploads/2020/08/Dorathy-1.jpg",
-    votes: 0
-  },
-  {
-    name: "Ozo",
-    votes: 0,
-    picture:
-      "https://scontent.fabv2-1.fna.fbcdn.net/v/t1.0-9/110161165_1150928051973512_5031557519512403774_o.jpg?_nc_cat=111&_nc_sid=dd9801&_nc_ohc=4FXuo0DJ2AEAX8avFNS&_nc_ht=scontent.fabv2-1.fna&oh=fdc7bc48961cf7042368df6738f6318d&oe=5F8A6493"
-  },
-  {
-    name: "Kiddwaya",
-    votes: 0,
-    picture:
-      "https://i1.wp.com/media.premiumtimesng.com/wp-content/files/2020/08/Screenshot-270-e1597078639510.png?fit=566%2C561&ssl=1"
-  },
-  {
-    name: "Laycon",
-    votes: 0,
-    picture:
-      "https://ichef.bbci.co.uk/news/1024/branded_pidgin/8C52/production/_113522953_oda11.jpg"
-  }
-];
+import staticHousemates from './HousematesList';
 
 const maximumVotes = 10;
